fix ops-nn_compile_ophost_warnings



# message auto-generated for no-merge-commit merge:
!127 merge master into master

fix ops-nn_compile_ophost_warnings

Created-by: jiangjiawei
Commit-by: jiangjiawei
Merged-by: turing_project1
Description: ## 描述
修改ophost编译告警

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
修复前
![image.png](https://raw.gitcode.com/user-images/assets/7665709/44f17ba6-67b6-4f1f-8500-c2229cbb52ac/image.png 'image.png')
修复后编译告警数量降低10个
![image.png](https://raw.gitcode.com/user-images/assets/7665709/40e3d5df-2b4c-4d08-879f-927889d28356/image.png 'image.png')
算子样例和UT测试
conv3d_backprop_input_v2
![image.png](https://raw.gitcode.com/user-images/assets/7665709/d9825fd1-9b2a-4a06-834a-19d700e5ae63/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/49ec7f70-2f80-4eac-baf1-25cedc53860a/image.png 'image.png')
apply_top_k_top_p_with_sorted
![image.png](https://raw.gitcode.com/user-images/assets/7665709/65d2e0de-6327-4985-bf66-1f9916aced53/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/1409d9bf-5dd0-41e1-b573-3d686d7d3d22/image.png 'image.png')
adaptive_avg_pool3d_grad
![image.png](https://raw.gitcode.com/user-images/assets/7665709/ae3cc8ff-d9e1-4287-9765-26d991dc9a4d/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/9a413494-7c59-4c05-b973-a5115f7fc506/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/c712da29-55ea-4c90-93d0-7d0767b11058/image.png 'image.png')
avg_pool3_d
![image.png](https://raw.gitcode.com/user-images/assets/7665709/664c1209-04d4-4488-adda-4ba163ee7274/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/927e2bfe-53a1-436a-97e0-c4932ff59e21/image.png 'image.png')
avg_pool3_d_grad
![image.png](https://raw.gitcode.com/user-images/assets/7665709/09f5fc34-5a1a-4c88-9963-cb56699cd717/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/451ec711-a22c-4133-950b-3386c6c18990/image.png 'image.png')
max_pool3d_with_argmax_v2
![image.png](https://raw.gitcode.com/user-images/assets/7665709/4b191ed3-eb38-45e2-beed-b4bf4421fb8c/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/efde9f39-cef7-49b8-b066-7d78873e6c7d/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/ff62fb39-5642-4365-9391-45645fc30790/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/ff8a9428-d43b-43a0-ae70-3e3fb7a711c1/image.png 'image.png')

## 文档更新
不涉及

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [x ] 其他，请描述：编译告警修复




See merge request: cann/ops-nn!127

+++ conv/conv3d_backprop_input_v2/op_host/op_tiling/arch35/conv3d_backprop_input_v2_inner_product_tiling.cpp
@@ -216,8 +216,9 @@ bool Conv3DDXV2InnerProductTiling::CheckVecTransEnable(
                            (runInfo_.kernel_d > 1 && runInfo_.dedx_d > 1);
     // 正常数据类型，baseM>=512时MMAD基本能掩盖MTE2
     bool isMmadCoverMte2 = l0Params.baseM >= BASIC_BLOCK_SIZE_512;
-    if (static_cast<int32_t>(dtypeByte_) == ge::GetSizeByDataType(ge::DT_BF16) && runInfo_.kernel_h <= NUM_THREE &&
-        runInfo_.kernel_w <= NUM_THREE) { // NUM_THREE = 3: loadToB1的效率与kernel大小有关
+    if (static_cast<int32_t>(dtypeByte_) == ge::GetSizeByDataType(ge::DT_BF16) &&
+        static_cast<uint32_t>(runInfo_.kernel_h) <= NUM_THREE &&
+        static_cast<uint32_t>(runInfo_.kernel_w) <= NUM_THREE) { // NUM_THREE = 3: loadToB1的效率与kernel大小有关
         // 16bit的loadToB1不需要transpose，MTE2带宽压力相对较小，实测baseM>=384时MMAD基本能掩盖MTE2
         isMmadCoverMte2 = l0Params.baseM >= (BASIC_BLOCK_SIZE_256 + BASIC_BLOCK_SIZE_128);
     }
